% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEI_annotation}
\alias{DEI_annotation}
\title{TAIR10 gene models from Arabidopsis_thaliana.TAIR10.47.gtf}
\format{
A data frame with 63 rows and 3 variables:
\describe{
  \item{SampleName}{Sample name}
  \item{Time}{Time point}
  \item{Treatment}{Photoperiod}
}
}
\usage{
DEI_annotation
}
\description{
TAIR10 gene models from Arabidopsis_thaliana.TAIR10.47.gtf
}
\keyword{datasets}
