% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edgeR_contrasts}
\alias{edgeR_contrasts}
\title{Contrasts used for identifying differentially expressed genes in the RNA-seq data}
\format{
A data frame with 108 observations and 7 variables:
\describe{
  \item{contrast}{Name of the contrast between individual time points}
  \item{cond1}{Photoperiod 1 of the contrast}
  \item{cond2}{Photoperiod 2 of the contrast}
  \item{time}{Time point}
  \item{sample_name}{Sample used in the contrast}
  \item{sample_cond}{Photoperiod of the sample}
  \item{contrast_conds}{Name of the contrast between photoperiods}
}
}
\usage{
edgeR_contrasts
}
\description{
Contrasts used for identifying differentially expressed genes in the RNA-seq data
}
\keyword{datasets}
