% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_get_DEGs.R
\name{edgeR_get_DEGs}
\alias{edgeR_get_DEGs}
\title{Extract significantly up- and down-reguluted genes from edgeR results}
\usage{
edgeR_get_DEGs(
  edgeR_result,
  grouping,
  sig_cutoff,
  FC_cutoff,
  PVAL_or_FDR = "PVAL"
)
}
\arguments{
\item{edgeR_result}{Data frame. Output from perform_edgeR.}

\item{grouping}{Factor. Generated from read_edgeR_contrasts}

\item{sig_cutoff}{Numerical.}

\item{FC_cutoff}{Numerical.}

\item{PVAL_or_FDR}{Character. "PVAL" or "FDR".}
}
\value{
None
}
\description{
Extract significantly up- and down-reguluted genes from edgeR results
}
