% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_MDSplot.R
\name{edgeR_MDSplot}
\alias{edgeR_MDSplot}
\title{Generate a MDS plot for gene count data}
\usage{
edgeR_MDSplot(gene_count_matrix)
}
\arguments{
\item{gene_count_matrix}{Matrix.}
}
\value{
ggplot2 object
}
\description{
Generate a MDS plot for gene count data
}
