% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_read_contrasts.R
\name{edgeR_read_contrasts}
\alias{edgeR_read_contrasts}
\title{Parse an edgeR contrast table into a list of list of contrasts and groupings.}
\usage{
edgeR_read_contrasts(edgeR_contrasts)
}
\arguments{
\item{edgeR_contrasts}{Dataframe. EdgeR contrasts}
}
\value{
None
}
\description{
Parse an edgeR contrast table into a list of list of contrasts and groupings.
}
