% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_run.R
\name{edgeR_run}
\alias{edgeR_run}
\title{Perform edgeR differentially expressed gene analysis}
\usage{
edgeR_run(raw_count_matrix, sample_names, grouping)
}
\arguments{
\item{raw_count_matrix}{Matrix. A numerical matrix with gene identifiers as row names and sample identifiers as column names.}

\item{sample_names}{Character. Sample identifiers for the analysis.}

\item{grouping}{Factor. A factor of numbers with the same length as sample_names to describe the grouping.}
}
\value{
Data frame
}
\description{
Perform edgeR differentially expressed gene analysis
}
