% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{photoperiodic_genes}
\alias{photoperiodic_genes}
\title{Photoperiodic genes in this dataset}
\format{
A vector of 8302 photoperiodic genes
}
\usage{
photoperiodic_genes
}
\description{
Photoperiodic genes in this dataset
}
\keyword{datasets}
