% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{read_count_matrix}
\alias{read_count_matrix}
\title{Contrasts used for identifying differentially expressed genes in the RNA-seq data}
\format{
A data frame with 108 observations and 7 variables:
\describe{
  \item{EQ_00_A}{Sample}
  \item{EQ_00_B}{Sample}
  \item{EQ_00_C}{Sample}
  \item{EQ_04_A}{Sample}
  \item{EQ_04_B}{Sample}
  \item{EQ_04_C}{Sample}
  \item{EQ_08_A}{Sample}
  \item{EQ_08_B}{Sample}
  \item{EQ_08_C}{Sample}
  \item{EQ_12_A}{Sample}
  \item{EQ_12_B}{Sample}
  \item{EQ_12_C}{Sample}
  \item{EQ_16_A}{Sample}
  \item{EQ_16_B}{Sample}
  \item{EQ_16_C}{Sample}
  \item{EQ_20_A}{Sample}
  \item{EQ_20_B}{Sample}
  \item{EQ_20_C}{Sample}
  \item{LD_00_A}{Sample}
  \item{LD_00_B}{Sample}
  \item{LD_00_C}{Sample}
  \item{LD_04_A}{Sample}
  \item{LD_04_B}{Sample}
  \item{LD_04_C}{Sample}
  \item{LD_08_A}{Sample}
  \item{LD_08_B}{Sample}
  \item{LD_08_C}{Sample}
  \item{LD_12_A}{Sample}
  \item{LD_12_B}{Sample}
  \item{LD_12_C}{Sample}
  \item{LD_16_A}{Sample}
  \item{LD_16_B}{Sample}
  \item{LD_16_C}{Sample}
  \item{LD_20_A}{Sample}
  \item{LD_20_B}{Sample}
  \item{LD_20_C}{Sample}
  \item{SD_00_A}{Sample}
  \item{SD_00_B}{Sample}
  \item{SD_00_C}{Sample}
  \item{SD_04_A}{Sample}
  \item{SD_04_B}{Sample}
  \item{SD_04_C}{Sample}
  \item{SD_08_A}{Sample}
  \item{SD_08_B}{Sample}
  \item{SD_08_C}{Sample}
  \item{SD_12_A}{Sample}
  \item{SD_12_B}{Sample}
  \item{SD_12_C}{Sample}
  \item{SD_16_A}{Sample}
  \item{SD_16_B}{Sample}
  \item{SD_16_C}{Sample}
  \item{SD_20_A}{Sample}
  \item{SD_20_B}{Sample}
  \item{SD_20_C}{Sample}
}
}
\usage{
read_count_matrix
}
\description{
Contrasts used for identifying differentially expressed genes in the RNA-seq data
}
\keyword{datasets}
