% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_edgeR.R
\name{run_edgeR}
\alias{run_edgeR}
\title{Extract significantly up- and down-reguluted genes from edgeR results}
\usage{
run_edgeR(
  dir,
  edgeR_contrast_table = edgeR_contrasts,
  gene_count_table = read_count_matrix,
  sig_cutoff = 0.2,
  FC_cutoff = 1
)
}
\arguments{
\item{dir}{Character. Directory.}

\item{edgeR_contrast_table}{Either a dataframe with edgeR contrasts, or full path to a tab-separated file of it}

\item{gene_count_table}{Either a dataframe with raw gene counts, or full path to a tab-separated file of it}

\item{sig_cutoff}{Numeric. p-value threshold for determining photoperiodic genes}

\item{FC_cutoff}{Numeric. Fold change threshold for determining photoperiodic genes}
}
\value{
None
}
\description{
Extract significantly up- and down-reguluted genes from edgeR results
}
