% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_input.R
\name{run_input}
\alias{run_input}
\title{Extract significantly up- and down-reguluted genes from edgeR results}
\usage{
run_input(
  dir,
  gene_count_table = read_count_matrix,
  edgeR_contrast_table = edgeR_contrasts,
  DEI_annotation_table = DEI_annotation
)
}
\arguments{
\item{dir}{Character. Directory.}

\item{gene_count_table}{Either a dataframe with raw gene counts, or full path to a tab-separated file of it}

\item{edgeR_contrast_table}{Either a dataframe with edgeR contrasts, or full path to a tab-separated file of it}

\item{DEI_annotation_table}{Either a dataframe with DEI annotation, or full path to a tab-separated file of it}
}
\value{
None
}
\description{
Extract significantly up- and down-reguluted genes from edgeR results
}
