% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_DEI.R
\name{ternary_DEI}
\alias{ternary_DEI}
\title{Plot ternary plot of gene DEI}
\usage{
ternary_DEI(
  power = 4,
  plot_density = FALSE,
  plot_contour = TRUE,
  plot_points = TRUE
)
}
\arguments{
\item{power}{Numeric.}

\item{plot_density}{Logical.}

\item{plot_contour}{Logical.}

\item{plot_points}{Logical.}
}
\value{
plot
}
\description{
Plot ternary plot of gene DEI
}
