% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TMM_mean_matrix}
\alias{TMM_mean_matrix}
\title{TAIR10 gene models from Arabidopsis_thaliana.TAIR10.47.gtf}
\format{
A data frame with 24845 rows and 18 variables:
\describe{
  \item{EQ_00}{Sample}
  \item{EQ_04}{Sample}
  \item{EQ_08}{Sample}
  \item{EQ_12}{Sample}
  \item{EQ_16}{Sample}
  \item{EQ_20}{Sample}
  \item{LD_00}{Sample}
  \item{LD_04}{Sample}
  \item{LD_08}{Sample}
  \item{LD_12}{Sample}
  \item{LD_16}{Sample}
  \item{LD_20}{Sample}
  \item{SD_00}{Sample}
  \item{SD_04}{Sample}
  \item{SD_08}{Sample}
  \item{SD_12}{Sample}
  \item{SD_16}{Sample}
  \item{SD_20}{Sample}
}
}
\usage{
TMM_mean_matrix
}
\description{
TAIR10 gene models from Arabidopsis_thaliana.TAIR10.47.gtf
}
\keyword{datasets}
